minor refactor of variables

+++ sqlg-test/src/main/java/org/umlg/sqlg/test/tp3/SqlgAbstractGraphProvider.java
@@ -18,7 +18,7 @@ import java.util.Set;
  */
 public abstract class SqlgAbstractGraphProvider extends AbstractGraphProvider {
 
-    protected static final Set<Class> IMPLEMENTATIONS = new HashSet<Class>() {{
+    private static final Set<Class> IMPLEMENTATIONS = new HashSet<Class>() {{
         add(SqlgEdge.class);
         add(SqlgElement.class);
         add(SqlgGraph.class);
@@ -32,41 +32,39 @@ public abstract class SqlgAbstractGraphProvider extends AbstractGraphProvider {
     public void clear(final Graph g, final Configuration configuration) throws Exception {
         SqlgDataSource sqlgDataSource = null;
         if (null != g) {
-            if (g.features().graph().supportsTransactions() && g.tx().isOpen())
+            if (g.features().graph().supportsTransactions() && g.tx().isOpen()) {
                 g.tx().rollback();
+            }
             g.close();
         }
         SqlDialect sqlDialect = getSqlgDialect(configuration);
         try {
-            sqlgDataSource = SqlgDataSource.setupDataSource(
-                    sqlDialect.getJdbcDriver(),
-                    configuration);
+            sqlgDataSource = SqlgDataSource.setupDataSource(sqlDialect.getJdbcDriver(), configuration);
             try (Connection conn = sqlgDataSource.get(configuration.getString("jdbc.url")).getConnection()) {
                 DatabaseMetaData metadata = conn.getMetaData();
                 if (sqlDialect.supportsCascade()) {
-                    String catalog = null;
-                    String schemaPattern = null;
                     String tableNamePattern = "%";
                     String[] types = {"TABLE"};
-                    ResultSet result = metadata.getTables(catalog, schemaPattern, tableNamePattern, types);
-                    while (result.next()) {
-                        StringBuilder sql = new StringBuilder("DROP TABLE ");
-                        sql.append(sqlDialect.maybeWrapInQoutes(result.getString(2)));
-                        sql.append(".");
-                        sql.append(sqlDialect.maybeWrapInQoutes(result.getString(3)));
-                        sql.append(" CASCADE");
-                        if (sqlDialect.needsSemicolon()) {
-                            sql.append(";");
-                        }
-                        try (PreparedStatement preparedStatement = conn.prepareStatement(sql.toString())) {
-                            preparedStatement.executeUpdate();
+                    ResultSet resultSet = metadata.getTables(null, null, tableNamePattern, types);
+                    while (resultSet.next()) {
+                        String schema = resultSet.getString(2);
+                        if (!sqlDialect.getGisSchemas().contains(schema)) {
+                            StringBuilder sql = new StringBuilder("DROP TABLE ");
+                            sql.append(sqlDialect.maybeWrapInQoutes(resultSet.getString(2)));
+                            sql.append(".");
+                            sql.append(sqlDialect.maybeWrapInQoutes(resultSet.getString(3)));
+                            sql.append(" CASCADE");
+                            if (sqlDialect.needsSemicolon()) {
+                                sql.append(";");
+                            }
+                            try (PreparedStatement preparedStatement = conn.prepareStatement(sql.toString())) {
+                                preparedStatement.executeUpdate();
+                            }
                         }
                     }
-                    catalog = null;
-                    schemaPattern = null;
-                    result = metadata.getSchemas(catalog, schemaPattern);
-                    while (result.next()) {
-                        String schema = result.getString(1);
+                    resultSet = metadata.getSchemas(null, null);
+                    while (resultSet.next()) {
+                        String schema = resultSet.getString(1);
                         if (!sqlDialect.getDefaultSchemas().contains(schema) && !sqlDialect.getGisSchemas().contains(schema)) {
                             StringBuilder sql = new StringBuilder("DROP SCHEMA ");
                             sql.append(sqlDialect.maybeWrapInQoutes(schema));
@@ -86,8 +84,9 @@ public abstract class SqlgAbstractGraphProvider extends AbstractGraphProvider {
         } catch (PropertyVetoException e) {
             throw new RuntimeException(e);
         } finally {
-            if (sqlgDataSource != null)
+            if (sqlgDataSource != null) {
                 sqlgDataSource.close(configuration.getString("jdbc.url"));
+            }
         }
     }
